tests: moved out fixture
